bump to latest, add mem64

+++ cpp/wllama.cpp
@@ -170,8 +170,8 @@ extern "C" const char *wllama_debug()
   {
 #ifdef __EMSCRIPTEN__
     auto i = mallinfo();
-    unsigned int total_mem = get_mem_total();
-    unsigned int dynamic_top = (unsigned int)sbrk(0);
+    size_t total_mem = get_mem_total();
+    size_t dynamic_top = (size_t)sbrk(0);
     return total_mem - dynamic_top + i.fordblks;
 #else
     return 0;
